# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/notifications/sms": {
            "get": {
                "description": "List Sms",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sms"
                ],
                "summary": "List Sms",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Sms"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "send sms",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SMS"
                ],
                "summary": "send sms",
                "parameters": [
                    {
                        "description": "SMS you want to create",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SmsRequestV1"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "The newly created SMS",
                        "schema": {
                            "$ref": "#/definitions/entity.Sms"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Sms": {
            "type": "object",
            "properties": {
                "activeStatus": {
                    "type": "boolean"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "00000000-0000-0000-0000-000000000000"
                },
                "messageId": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "request.SmsRequestV1": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        }
    }
}